[task2pomdp] Adds end reward and failure/no probabilities.
These relax the POMDP model but also improve the solution found by the
POMDP solver. The fact that the 'end' state was only rewarded by
enabling no cost on 'wait' action was not sufficient for the solver
which seems to get blinded by short term high costs for acting and often
chose to wait instead. The end reward improves on this. Probabilities of
failure of robot action (even when taken in correct context) as well as
human refusing to act as robot suggested also seem to improve the final
policy by making the solver better explore around the default path.