WIP: parallel branch to POMDP.
This is implemented by converting parallel into a sequence of
alternatives. The current implementation does not duplicate and rename
the children nodes and action so it leaves a tree with duplicated
actions which the POMDP converter does not support. Action and node deep
copy with rename needs to be implemented for it to be compatible.